#1 i2c working with write only (i.e. no reading)
i2c only writes and does not read from the LCD screen. This could be  a screen problem or a i2c bus fault I do not know. Only things that were read were the chip id which states that the screen is working and the ring buffer position which does not need to be done for this project.